kirisame/requirements.txt
django==2.1.2  # Web framework
redis==2.10.6  # Redis driver
psycopg2-binary==2.7.4  # PostgreSQL driver

git+git://github.com/celery/celery.git  # Task queue manager
django-celery-beat==1.1.1  # Scheduler for periodic tasks
django-celery-results==1.0.1  # View task results in Django
gevent==1.3.7  # Concurrency pool manager

requests==2.19.1  # HTTP requests for humans
django-fsm==2.6.0  # Finite state machine for Django models
django-constance[database]==2.2.0  # Settings in admin site

pixivpy==3.3.4  # Pixiv API
Mastodon.py==1.3.1  # Mastodon API

opencv-python==3.4.3.18  # For face recognition
